Splits editor components into editor-add/edit
Wires components according

Deletes edit or add component actions depending on the component

Deletes if statements previously swapping submit buttons based on edit/
add

closes #61
