+++ app/templates/invadd.hbs
@@ -1,1 +1,1 @@
-{{inv-editor createEntry=createEntry}}
+{{inv-editor-add}}
