+++ app/templates/invedit.hbs
@@ -1,1 +1,1 @@
-{{inv-editor}}
+{{inv-editor-edit}}
