+++ app/templates/wladd.hbs
@@ -1,1 +1,1 @@
-{{wl-editor createEntry=createEntry}}
+{{wl-editor-add}}
