+++ app/templates/wledit.hbs
@@ -1,1 +1,1 @@
-{{wl-editor}}
+{{wl-editor-edit}}
